+++ java-core/src/main/java/org/lessons/lesson03/Car.java
@@ -1,0 +1,20 @@
+package org.lessons.lesson03;
+
+public class Car {
+    String country;
+    String model;
+    int yearOfManufacture;
+
+    public Car(String country, String model, int yearOfManufacture) {
+        this.country = country;
+        this.model = model;
+        this.yearOfManufacture = yearOfManufacture;
+    }
+
+    public void printCarInfo(){
+        System.out.println("Модель: " + model +
+            ", страна производства: " + country +
+            ", год выпуска: " + yearOfManufacture);
+    }
+
+}
