+++ java-core/src/main/java/org/lessons/lesson03/Main.java
@@ -29,70 +29,4 @@ public class Main {
         car1.printCarInfo();
     }
 }
-// 1 task
-class Person {
-    String name;
-    int age;
 
-    public Person(String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public void introduce() {
-        System.out.println("Здравствуйте! Меня зовут " + name + ", мне " + age + " лет");
-    }
-}
-
-// 2 task
-class Rectangle{
-    double length;
-    double width;
-
-    public Rectangle(double length, double width) {
-        this.length = length;
-        this.width = width;
-    }
-
-    public double CalculateSquare(){
-        return  length * width;
-    }
-
-    public double CalculatePerimeter(){
-        return 2 * (length + width);
-    }
-}
-
-// 3 task
-class Car {
-    String country;
-    String model;
-    int yearOfManufacture;
-
-    public Car(String country, String model, int yearOfManufacture) {
-        this.country = country;
-        this.model = model;
-        this.yearOfManufacture = yearOfManufacture;
-    }
-
-    public void printCarInfo(){
-        System.out.println("Модель: " + model +
-            ", страна производства: " + country +
-            ", год выпуска: " + yearOfManufacture);
-    }
-
-}
-
-//4 task
-class BankAccount{
-    String fio;
-    int AccountNumber;
-    double balance;
-
-    public BankAccount(String fio, int AccountNumber, double balance) {
-        this.fio = fio;
-        this.AccountNumber = AccountNumber;
-        this.balance = balance;
-
-    }
-}
